+++ src/api/user/index.ts
@@ -6,7 +6,6 @@ import {
   getAllUserHandler,
   getUserHandler,
   updateUserHandler,
-  loginHandler,
 } from './user.controller';
 
 const router = Router();
@@ -25,7 +24,5 @@ router.delete('/:id', deleteUserHandler);
 
 // /api/users/:id -> PATCH
 router.patch('/:id', updateUserHandler);
-
-router.post('/login', loginHandler);
 
 export default router;
