+++ src/api/user/user.controller.ts
@@ -71,26 +71,3 @@ export async function updateUserHandler(req: Request, res: Response) {
 
   return res.json(updatedUser);
 }
-
-export async function loginHandler(req: Request, res: Response) {
-  const { email, password } = req.body;
-
-  try {
-    const user = await getUserByEmail(email);
-
-    if (!user) {
-      return res.status(404).json({ message: 'User not found' });
-    }
-
-    // compare password
-    const isMatch = await comparePassword(password, user.password);
-
-    if (!isMatch) {
-      return res.status(400).json({ message: 'Email or password not match' });
-    }
-
-    // jwt
-
-    return res.json(user);
-  } catch (error) {}
-}
